refactoring webhooks to separate platform from connect
